requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.6.4
MarkupSafe==2.0.1
multitasking==0.0.10
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
urllib3==1.26.7
Werkzeug==2.0.2
yfinance==0.1.67
